// src/presentation/http/customers/delete-customer.dto.ts
import { ApiProperty } from '@nestjs/swagger';
import { IsEmail, IsUUID, Length } from 'class-validator';
import { IdentificationType, CustomerAddress } from '@prisma/client';

export namespace DeleteCustomerDto {
  export class Request {
    @ApiProperty({
      required: true,
      example: '057e0255-e3fa-4a6f-8101-510c8e6b060b',
      description: 'Product Customer ID - uuid'
    })
    @IsUUID()
    id: string;
  }
  export class Response {
    id: string;
    user_id: string;
    identification: number;
    identification_type: IdentificationType;
    customer_addresses?: CustomerAddress[];
    created_at: Date;
    updated_at: Date;
    deleted_at: Date;
  }
}